+++ source/data.js
@@ -1,5 +1,5 @@
 // builtin
-import { cwd } from 'process'
+import { cwd } from 'node:process'
 
 // external
 import {